Relocate LAMMPS skill + tools to upstream Tier C layout
Upstream standardized on a Tier C bundle layout while this branch
was off in the wilderness:

  scilink/skills/<category>/<name>/<name>.md   (skill markdown)
  scilink/skills/<category>/<name>/<name>.py   (per-skill helper, optional)
  scilink/skills/<category>/<name>/__init__.py
  scilink/tools/                                (mostly gone -- tools
                                                 now live with their skill,
                                                 cf. force_field/amber/amber.py)

Move:
  scilink/skills/molecular_dynamics/lammps.md
    -> scilink/skills/molecular_dynamics/lammps/lammps.md
  scilink/tools/molecular_dynamics/lammps_tools.py
    -> scilink/skills/molecular_dynamics/lammps/lammps.py

Add per-bundle __init__.py files so the loader's path resolution
finds the bundles.

Update import sites:
  md_simulation_agent.py:
    from ...skills.molecular_dynamics.lammps import lammps as lammps_tools
  tests/test_lammps_skill/test_lammps_tools.py: same import update

Also fix a typo introduced in 'Refactored, generalized MD simulation
agent' (the cherry-picked equivalent of upstream's 9f24c5f) where
md_simulation_agent.py imported `.base_simulation_agent` instead of
`.base_agent`. The typo was fixed later in user's mlip_agents history
inside the MLIP commit (which lives on a separate branch / future
PR), so this branch needs the fix included here.

Expose the new agents in sim_agents/__init__.py:
  SimulationAgent (base), MDSimulationAgent.
LAMMPSSimulationAgent stays as the backward-compat wrapper.

This is a pure layout move + one-line typo fix; no functional
changes. 119 tests pass under tests/test_lammps_skill/ +
tests/test_skill_loader.py.

The architectural intent (generic MD agent + per-package skills and
tools) is now realized in code: the agent has no LAMMPS-specific
methods; LAMMPS-specific behavior comes from the skill markdown +
the LAMMPS tool module discovered via TOOL_REGISTRY. Adding GROMACS
or OpenMM is now "drop in a skill bundle + a tools module + register
it" rather than "subclass LAMMPSSimulationAgent."
